requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
coloredlogs==14.0         # via brightsky (setup.py)
cssselect==1.1.0          # via parsel
humanfriendly==8.1        # via coloredlogs
idna==2.9                 # via requests
lxml==4.5.0               # via parsel
parsel==1.5.2             # via brightsky (setup.py)
python-dateutil==2.8.1    # via brightsky (setup.py)
requests==2.23.0          # via brightsky (setup.py)
six==1.14.0               # via parsel, python-dateutil, w3lib
urllib3==1.25.8           # via requests
w3lib==1.21.0             # via parsel
